Add Sort button, input styles
Initiative can now be sorted using a sort button. The sorted order is set to and read from the scene.

The initiative input can be selected with a single click and has hover and selection effects.

Long words for names are now wrapped properly.

Added back the explicit hidden icon.